requirements.txt
gdown>=5.2.0
networkx>=3.3
shapely>=2.0.4
opencv-python>=4.10.0.84
gymnasium>=0.29.1
pygame>=2.6.0
scikit-image>=0.24.0
matplotlib>=3.9.0
joblib>=1.4.2
moviepy>=1.0.3
python-dotenv>=1.0.1
openai>=1.39.0
pandas>=2.2.2
seaborn>=0.13.2
pycocotools>=2.0.8
colorama>=0.4.6
tabulate==0.9.0
scenic=="2.1.0b4
git+https://github.com/ymd-h/cpprb.git